requirements.txt
ExifRead==2.3.2
PyYAML==6.0
